Organized the project directory
